requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
alembic==0.8.10           # via dataset
appdirs==1.4.0            # via setuptools
boto3==1.4.4
botocore==1.5.7           # via boto3, s3transfer
cchardet==1.1.2           # via tabulator
chardet==2.3.0            # via normality
click==6.7                # via jsontableschema, python-dotenv, tabulator
contextlib2==0.5.4        # via raven
datapackage==0.6.1
dataset==0.7.1
docutils==0.13.1          # via botocore
et-xmlfile==1.0.1         # via openpyxl
ezodf==0.3.2              # via tabulator
functools32==3.2.3.post2  # via jsonschema
future==0.16.0            # via jsontableschema
futures==3.0.5            # via s3transfer
fuzzywuzzy==0.14.0
ijson==2.3                # via tabulator
iso3166==0.7
jdcal==1.3                # via openpyxl
jmespath==0.9.1           # via boto3, botocore
jsonlines==1.1.0          # via tabulator
jsonschema==2.5.1         # via datapackage, jsontableschema
jsontableschema-sql==0.5.0
jsontableschema==0.9.0    # via datapackage, jsontableschema-sql
linear-tsv==1.0.0         # via tabulator
lxml==3.7.2               # via tabulator
mako==1.0.6               # via alembic
markupsafe==0.23          # via mako
normality==0.3.9          # via dataset
openpyxl==2.4.2           # via tabulator
packaging==16.8           # via setuptools
psycopg2==2.6.2
pybossa-client==1.1.1
pyparsing==2.1.10         # via packaging
pypdf2==1.26.0
python-dateutil==2.6.0    # via botocore, jsontableschema, python-documentcloud
python-documentcloud==1.0.4
python-dotenv==0.6.2
python-editor==1.0.3      # via alembic
python-levenshtein==0.12.0
pyyaml==3.12              # via dataset
raven==5.32.0
requests==2.13.0
rfc3986==0.4.1            # via jsontableschema
rfc3987==1.3.7            # via python-documentcloud
s3transfer==0.1.10        # via boto3
six==1.10.0               # via datapackage, dataset, jsonlines, jsontableschema-sql, linear-tsv, normality, packaging, python-dateutil, python-documentcloud, setuptools, tabulator
sqlalchemy==1.1.5
tabulator==0.14.0         # via datapackage, jsontableschema, jsontableschema-sql
unicodecsv==0.14.1        # via datapackage, jsontableschema, tabulator
xlrd==1.0.0               # via tabulator

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via python-levenshtein
